+++ src/features/auth/forms/ForgotPasswordForm/ForgotPasswordForm.tsx
@@ -6,7 +6,7 @@ import { useForm } from 'react-hook-form'
 import { z } from 'zod'
 import { zodResolver } from '@hookform/resolvers/zod'
 import Link from 'next/link'
-import { Recaptcha } from '@/entities/Recaptcha'
+import { Recaptcha } from '@/shared/ui/Recaptcha'
 import { PublicRoutes } from '@/shared/enums'
 import { ControlledInput } from '@/shared/components/Controlled/ControlledInput'
 import { useState } from 'react'
