+++ src/shared/ui/Recaptcha/Recaptcha.stories.tsx
@@ -1,4 +1,4 @@
-import { Recaptcha } from '@/entities/Recaptcha/Recaptcha'
+import { Recaptcha } from '@/shared/ui/Recaptcha/Recaptcha'
 import { store } from '@/shared/store'
 import { Meta, StoryObj } from '@storybook/nextjs-vite'
 import { Provider } from 'react-redux'
