+++ src/shared/ui/Recaptcha/RecaptchaSpinner.tsx
